backend_fastapi/requirements.txt
black==22.10.0
fastapi==0.88.0
flake8==6.0.0
mysql-connector-python==8.0.31
python-dotenv==0.21.0
uvicorn==0.20.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5